BroadcastReceiver вынесен из WeatherPresenter
